pr-preview/pr-67/recipe_automation/requirements.txt
gspread
oauth2client
python-slugify
GitPython
requests
google-api-python-client
google-auth
Pillow
